Создание Фрагментов, реализация добавления друзей в список и передача списка между Активити и Фрагментами задача 12_02
