+++ task_12/valera/app/src/main/java/com/example/myapplication/Friend.kt
@@ -1,7 +1,11 @@
 package com.example.myapplication
 
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
 data class Friend(
     val name: String,
     val surname: String,
     val phoneNumber: String
-)
+) : Parcelable
