+++ task_12/valera/app/src/main/res/layout/activity_main.xml
@@ -34,69 +34,16 @@
             android:layout_height="wrap_content"
             android:orientation="vertical">
 
-            <TextView
-                android:id="@+id/welcome_words"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center_vertical"
-                android:gravity="center"
-                android:text="@string/welcome_words"
-                android:textSize="25sp" />
-            <LinearLayout
-                android:id="@+id/adder_friend"
+            <androidx.fragment.app.FragmentContainerView
+                android:id="@+id/fragment_container"
+                android:name="com.example.myapplication.FragmentWelcomeWords"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/toolbar"
-                android:visibility="gone"
-                android:orientation="vertical">
+                tools:layout="@layout/fragment_adder_friends">
+            </androidx.fragment.app.FragmentContainerView>
 
-                <EditText
-                    android:id="@+id/edit_name"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:hint="@string/name"
-                    android:inputType="textPersonName"
-                    />
-                <EditText
-                    android:id="@+id/edin_surname"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:hint="@string/surname"
-                    android:inputType="textPersonName"
-                    />
-                <EditText
-                    android:id="@+id/edit_phone_number"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:hint="@string/phone_number"
-                    android:inputType="phone"
-                    />
-
-                <Button
-                    android:id="@+id/adder_buttom"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_gravity="center"
-                    android:text="@string/buttom_adder"
-                    android:enabled="false"
-                    />
-                <ProgressBar
-                    android:id="@+id/progressBar"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:visibility="gone"
-                    android:layout_gravity="center"
-                    />
-            </LinearLayout>
-
-            <LinearLayout
-                android:id="@+id/shower_list"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical"
-                android:visibility="gone"
-                />
             <LinearLayout
                 android:id="@+id/searcher_liner_layout"
                 android:layout_width="match_parent"
